feat(backend): actualización de controladores, modelos y migraciones
- Modifica DatabaseController y ReporteDocenteController
- Cambios en modelo Materia
- Eliminación de modelo Plan
- Actualización de migraciones (materias, planes, grupos)
- Ajustes en rutas API
- Cambios en composer.json y composer.lock
- Nueva migración migration_snapshots

+++ backend/database/migrations/2026_04_20_130041_create_materias_table.php
@@ -8,7 +8,7 @@ return new class extends Migration {
     public function up(): void
     {
         Schema::create('materias', function (Blueprint $table) {
-            $table->decimal('anio', 10, 0);
+            $table->integer('anio');
             $table->string('plan', 6);
             $table->string('codigo', 7);
             $table->string('sigla', 45)->nullable();
@@ -18,8 +18,8 @@ return new class extends Migration {
             $table->string('obligatorio', 1)->nullable();
             $table->string('tipo', 40);
             $table->string('activo', 1)->nullable();
-            $table->decimal('departamento', 10, 0)->nullable();
-            // Sin primary key — tiene duplicados históricos intencionales
+            $table->integer('departamento')->nullable();
+            
         });
     }
